Extensive new checking for arity
The arities of all "concrete" types are verified.
A concrete type is one used with a value: argument types, return types,
and variable/field types. Their generic arguments are non-concrete.

The reason for this is higher-order types:

class MaybeCollection[#C, #T] {
public field fin #C[Maybe[#T]] : myTable;
}

If we instantiate a MaybeCollection[List, Int] we might erroneously
conclude that `List` is arity-0 since it has no parameters here; in
reality, it is given one by the 'maybecollection' class.

In the future, we may also support 'partial application with currying'
for types, which would allow you to do things like this:

class HashMap[#Key, #Value] {
// elided for brevity
}

MaybeCollection[ HashMap[String], Bool ] : myThing;

which gives incomplete type HashMap[String] as a parameter, with a
"hole" that must be filled. If we inspect the type of our 'myTable' for
this instance, we see by simple substitution that it ought to be:

#C[Maybe[#T]]

Where #C = HashMap[String] and #T = Bool. Thus:

HashMap[String][Maybe[Bool]]

We see this strange T[X][Y] pattern; we interpret it to mean T[X, Y]
(uniquely and consistently) so that our object has type:

HashMap[ String, Maybe[Bool] ]

Which is a natural composition of types.

There's also another possible direction, of 'holes' in types:

MaybeCollection[ HashMap[ _ , Int ], String ] : myThing;

Which will have a 'myTable' type of:
HashMap[ Maybe[String], Int ]

The usefulness/safety of this design will have to be weighed at a later
date.