Eliminated redundant 'Locate' wrapper around 'Type'
In Parse.Type, we can see that prior to this commit, we have the type
definitions:

data FieldType
= FieldType { typeName :: Name, typeArguments :: [Type] }
deriving Show
type Type = Locate FieldType

However, this is pointless, since the 'typeName' will always point to
the same place as the 'Locate'.

The reason for this redundancy is that originally our types were:

data FieldType
= FieldType { typeName :: String, typeArguments :: [Type] }
deriving Show
type Type = Locate FieldType

Before I had thought of the "Name" alias for "Locate String".

Thus we finally fix this headache. The new, simpler definition is:

data Type
= Type { typeName :: Name, typeArguments :: [Type] }

Foregoing our previous type synonym.

However, we lose the convenience of being able to use 'at' to locate
types so we introduce a new function called 'typeAt' that tells you its
location.