Remove protected; check field visibility
This change removes the "protected" special keyword.

It also incorporated checking visibility into the field-get routine, so
that private fields cannot be accessed by outside classes.

This change affects the most general field-get interface (through
Environ), so it will propagate through all future code which attempts to
get fields.

Later work will implement the same for methods and constructors.

--

Protected was removed due to the decision to remove inheritance from the
language. It's possible that this will persist in some different form,
but it's not yet clear how it will work. For now, only "interfaces"
(having more similarity to Haskell typeclasses than Java-style
interfaces) will be the means for polymorphic many-behavior functions.