Removed "discard" from language
This change makes verification of effects much, much easier.

In simple terms, discarding a reference is a "must" behavior, while all
other effects are just "might" behaviors. Requiring that it occur is
much harder to check than considering the possibility.

This simplification greatly improves the ability to develop further. In
the future, it may be reincorporated to the language in the same or
different form.