Fixed bugs in Flow and made it simpler
Previously, the following block would have failed to compile:

if (condition) {
break;
} else {
return;
}

The reason is that the encoding could only say that the code "might"
return and "might" break, but it could not say that at least one of them
had to occur.

The new way includes three parameters (and Bool instead of Flow):

mayProceed - it is (statically) possible to execute following code
mayReturn - it is statically possible to have a return in here
mayBreak - it is statically possible to have a break in here

A couple (currently undocumented) notes:

"while (true)" is not given any special significance. As such, code
after the loop, even when it contains no breaks, will always be
considered reachable.

As an alternative, a "forever" loop will be created (although the
presence of a break will cause the code after to become reachable).

In addition, it's illegal for a while-loop to end with unreachable code.
For example,

while (condition) {
if (condition2) {
return;
} else {
break;
}
}

The above will fail to compile, since the while's end is unreachable
(and thus it acts like an "if").

The inclusion of "continue" may cause this to change. To rectify this,
one of the following behaviors (as far as I can see) will need to be
chosen:
- Allow while-loops to have unreachable ends when they contain a
reachable continue statement
- Allow if-statements to use "continue" (restarting at the condition,
for consistency)
- Remove the requirement which prevents while-loops from being written
this way in its entirety